+++ zio-http/src/main/scala/zhttp/http/middleware/package.scala
@@ -1,0 +1,6 @@
+package zhttp.http
+
+package object middleware {
+  type HttpMiddleware[-R, +E] = Middleware[R, E, Request, Response, Request, Response]
+  type RequestP[+A]           = (Method, URL, Headers) => A
+}
