+++ zio-http/src/test/scala/zhttp/http/middleware/MiddlewareSpec.scala
@@ -1,6 +1,6 @@
-package zhttp.middleware
+package zhttp.http.middleware
 
-import zhttp.http.{HExitAssertion, Http, Middleware}
+import zhttp.http.{HExitAssertion, Http}
 import zio.duration._
 import zio.test.Assertion.{equalTo, isLeft, isNone, isSome}
 import zio.test.environment.TestClock
@@ -18,48 +18,48 @@ object MiddlewareSpec extends DefaultRunnableSpec with HExitAssertion {
     } +
       testM("codec") {
         val mid = Middleware.codec[String, Int](decoder = _.toInt, encoder = _.toString)
-        val app = Http.identity[Int] @@@ mid
+        val app = Http.identity[Int] @@ mid
         assertM(app("1"))(equalTo("1"))
       } +
       testM("constant") {
         val mid = Middleware.fromHttp(Http.succeed("OK"))
-        val app = Http.succeed(1) @@@ mid
+        val app = Http.succeed(1) @@ mid
         assertM(app(()))(equalTo("OK"))
       } +
       testM("interceptZIO") {
         for {
           ref <- Ref.make(0)
           mid = Middleware.interceptZIO[Int, Int](i => UIO(i * 10))((i, j) => ref.set(i + j))
-          app = Http.identity[Int] @@@ mid
+          app = Http.identity[Int] @@ mid
           _ <- app(1)
           i <- ref.get
         } yield assert(i)(equalTo(11))
       } +
       testM("orElse") {
         val mid = Middleware.fail("left") <> Middleware.fail("right")
-        val app = Http.empty @@@ mid
+        val app = Http.empty @@ mid
         assertM(app(()).flip)(isSome(equalTo("right")))
       } +
       testM("combine") {
         val mid1 = increment
         val mid2 = increment
         val mid  = mid1 compose mid2
-        val app  = Http.identity[Int] @@@ mid
+        val app  = Http.identity[Int] @@ mid
         assertM(app(0))(equalTo(4))
       } +
       testM("flatMap") {
         val mid = increment.flatMap(i => Middleware.succeed(i + 1))
-        val app = Http.identity[Int] @@@ mid
+        val app = Http.identity[Int] @@ mid
         assertM(app(0))(equalTo(3))
       } +
       testM("mapZIO") {
         val mid = increment.mapZIO(i => UIO(i + 1))
-        val app = Http.identity[Int] @@@ mid
+        val app = Http.identity[Int] @@ mid
         assertM(app(0))(equalTo(3))
       } +
       testM("race") {
         val mid = Middleware.succeed('A').delay(2 second) race Middleware.succeed("B").delay(1 second)
-        val app = Http.succeed(1) @@@ mid
+        val app = Http.succeed(1) @@ mid
         assertM(app(()) <& TestClock.adjust(3 second))(equalTo("B"))
       } +
       suite("ifThenElse") {
@@ -68,11 +68,11 @@ object MiddlewareSpec extends DefaultRunnableSpec with HExitAssertion {
           isFalse = i => Middleware.succeed(i - 1),
         )
         testM("isTrue") {
-          val app = Http.identity[Int] @@@ mid
+          val app = Http.identity[Int] @@ mid
           assertM(app(10))(equalTo(11))
         } +
           testM("isFalse") {
-            val app = Http.identity[Int] @@@ mid
+            val app = Http.identity[Int] @@ mid
             assertM(app(1))(equalTo(0))
           }
       } +
@@ -82,11 +82,11 @@ object MiddlewareSpec extends DefaultRunnableSpec with HExitAssertion {
           isFalse = i => Middleware.succeed(i - 1),
         )
         testM("isTrue") {
-          val app = Http.identity[Int] @@@ mid
+          val app = Http.identity[Int] @@ mid
           assertM(app(10))(equalTo(11))
         } +
           testM("isFalse") {
-            val app = Http.identity[Int] @@@ mid
+            val app = Http.identity[Int] @@ mid
             assertM(app(1))(equalTo(0))
           }
       }
